Add targetclass and move Coordinate to seperate file
