--- gym-drill/videos/openaigym.episode_batch.0.8484.stats.json
@@ -1,1 +1,0 @@
-{"initial_reset_timestamp": 1593030174.688367, "timestamps": [], "episode_lengths": [], "episode_rewards": [], "episode_types": ["t"]}
