--- gym-drill/videos/openaigym.manifest.0.8484.manifest.json
@@ -1,1 +1,0 @@
-{"stats": "openaigym.episode_batch.0.8484.stats.json", "videos": [["openaigym.video.0.8484.video000000.mp4", "openaigym.video.0.8484.video000000.meta.json"]], "env_info": {"gym_version": "0.17.2", "env_id": "drill-v0"}}
