--- gym-drill/videos/openaigym.video.0.8484.video000000.meta.json
@@ -1,1 +1,0 @@
-{"episode_id": 0, "content_type": "video/mp4", "empty": true}
