--- temp/temp.py
@@ -1,12 +1,0 @@
-import matplotlib.pyplot as plt
-import numpy as np
-
-
-print("heihei")
-
-x = np.linspace(0,10,10)
-
-plt.plot(x)
-
-plt.show()
-
