--- videos/openaigym.episode_batch.0.12140.stats.json
@@ -1,1 +1,0 @@
-{"initial_reset_timestamp": 1593032759.9288135, "timestamps": [], "episode_lengths": [], "episode_rewards": [], "episode_types": ["t"]}
