--- videos/openaigym.manifest.0.12140.manifest.json
@@ -1,1 +1,0 @@
-{"stats": "openaigym.episode_batch.0.12140.stats.json", "videos": [["openaigym.video.0.12140.video000000.mp4", "openaigym.video.0.12140.video000000.meta.json"]], "env_info": {"gym_version": "0.17.2", "env_id": "drill-v0"}}
