--- videos/openaigym.video.0.12140.video000000.meta.json
@@ -1,1 +1,0 @@
-{"episode_id": 0, "content_type": "video/mp4", "empty": true}
